requirements.txt
fastapi
uvicorn
pandas
matplotlib
seaborn
python-multipart
google-genai
reportlab
n8n